pre-release adding link support
Consider this a dead-end branch. Using Form and Profile as cornerstones for
navigating links was in hindsight an unrealistic choice, as they are not
widely used by many APIs in production. Middleware is arguably overengineered
as well, and could probably be simplified. Finally, some important tasks such
as consistent formatting for headers, link resolution utilities, testing, and
docs are not yet complete.

I'm throwing this up on Github anyway, mainly to close this chapter of
development and move forward with v1.0.

- TODO gut form/profile reliance
- TODO simplify middleware utilities
- TODO better test coverage
- TODO well-defined header format
- TODO infrastructure for resolving relative link URIs
- TODO documentation